requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic>=2.9.0
pymongo==4.6.0
requests==2.31.0
email-validator==2.1.0
argon2-cffi==23.1.0
python-multipart==0.0.9
itsdangerous==2.2.0
